+++ capstone/samples/simple.c
@@ -5,7 +5,6 @@
 #include <capstone/platform.h>
 #include <capstone/capstone.h>
 
-
 void print_string_hex(FILE *fp, const char *comment, unsigned char *str,
 		      size_t len)
 {
@@ -60,6 +59,26 @@ int main(void)
 #define X86_64_CODE \
 	"\x55\x48\x8b\x05\xb8\x13\x00\x00\xe9\xea\xbe\xad\xde\xff\x25\x23\x01\x00\x00\xe8\xdf\xbe\xad\xde\x74\xff"
 
+/* see kvm/playground/load-payload01-program.asm */
+#define X86_32_CODE            \
+	"\x66\xba\xf8\x03"     \
+	"\xb8\x61\x62\x63\x64" \
+	"\xe8\x01\x00\x00\x00" \
+	"\xf4"                 \
+	"\x52"                 \
+	"\x51"                 \
+	"\xb9\x04\x00\x00\x00" \
+	"\x66\xba\xf8\x03"     \
+	"\xee"                 \
+	"\xc1\xe8\x08"         \
+	"\x83\xe9\x01"         \
+	"\x75\xf7"             \
+	"\xb0\x78"             \
+	"\xee"                 \
+	"\x59"                 \
+	"\x5a"                 \
+	"\xc3"
+
 #define AArch64_CODE \
 	"\x09\x00\x38\xd5" \
 	"\xbf\x40\x00\xd5" \
@@ -90,6 +109,13 @@ int main(void)
 	mode = CS_MODE_64;
 	code = (unsigned char *)X86_64_CODE;
 	size = sizeof(X86_64_CODE) - 1;
+	print_string_hex(stdout, "Code:", code, size);
+	fdisasm(stdout, arch, mode, code, size);
+
+	arch = CS_ARCH_X86;
+	mode = CS_MODE_32;
+	code = (unsigned char *)X86_32_CODE;
+	size = sizeof(X86_32_CODE) - 1;
 	print_string_hex(stdout, "Code:", code, size);
 	fdisasm(stdout, arch, mode, code, size);