+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- scripts/version: add capstone.sh;
 #### Changed
 - kvm/playground: rename common.c to `kvm_helpers.c`;
 #### Deprecated